improve casts

+++ platform/default/src/mbgl/util/compression.cpp
@@ -32,16 +32,16 @@ namespace util {
 #undef compress
 
 bool is_compressed(const std::string &v) {
-    if ((uint8_t)v[0] == 0x1f && (uint8_t)v[1] == 0x8b) {
+    if (static_cast<uint8_t>(v[0]) == 0x1f && static_cast<uint8_t>(v[1]) == 0x8b) {
         // gzip (rfc1952)
         return true;
-    } else if ((uint8_t)v[0] == 0x78) {
+    } else if (static_cast<uint8_t>(v[0]) == 0x78) {
         // zlib (rfc1950)
-        switch ((uint8_t)v[1]) {
-            case (uint8_t)0x01: // 78 01 - No Compression/low
-            case (uint8_t)0x5E: // 78 5E - Fast Compression
-            case (uint8_t)0x9C: // 78 9C - Default Compression
-            case (uint8_t)0xDA: // 78 DA - Best Compression
+        switch (v[1]) {
+            case 0x01: // 78 01 - No Compression/low
+            case 0x5E: // 78 5E - Fast Compression
+            case 0x9C: // 78 9C - Default Compression
+            case 0xDA: // 78 DA - Best Compression
                 return true;
             default:
                 return false;
